Fixed build break.
[CL 30803848 by zach bethel in ue5-main branch]

+++ Engine/Source/Runtime/RenderCore/Public/RenderGraphAllocator.h
@@ -124,7 +124,7 @@ public:
 private:
 #if RDG_USE_MALLOC
 	TArray<void*> Mallocs;
-	uint32 NumMallocBytes = 0;
+	uint64 NumMallocBytes = 0;
 #else
 	FMemStackBase MemStack;
 #endif
